pyproject.toml
[tool.poetry]
name = "vfs-global-bot"
version = "0.1.0"
description = "VFS Global bot"
authors = []

[tool.isort]
default_section="THIRDPARTY"
indent='    '
known_tests="pytest"
sections=["FUTURE","STDLIB","THIRDPARTY","FIRSTPARTY","TESTS","LOCALFOLDER"]
multi_line_output=3
line_length=120
include_trailing_comma=true
use_parentheses=true
lines_after_imports=2

[tool.black]
line-length = 119
skip-string-normalization = true

[tool.poetry.dependencies]
python = "3.10.6"

numpy = "*"
opencv-python = "*"
pydantic = "*"
python-telegram-bot = "*"
selenium = "*"
selenium-recaptcha-solver = "*"
undetected-chromedriver = "*"

[tool.poetry.dev-dependencies]
black = "22.3.0"
isort = "*"

[build-system]
requires = ["poetry>=1.1"]
build-backend = "poetry.core.masonry.api"
